First commit of working resource provider create
If we run `python3 placement/__init.py <URL>` against a docker
container using placedock with

    docker run -t -p 127.0.0.1:8081:80 --env-file dockerenv cdent/placedock

500 random resource providers will be created.

The nesting of the various coroutines and the use of semaphores
is required to overcome the docker for mac proxy not being super
awesome.

Future commits will add inventory and nested-ness to resources being
created. Then allocation candidates will be requested, and allocations
written.

The idea here is to do much of that in vaguely concurrent fashions to
see how a placement service reacts.

[1] https://hub.docker.com/r/cdent/placedock/